pyproject.toml
[tool.poetry]
name = "hyde"
version = "0.1.0"
description = ""
authors = ["Johannes Windelen <mail@johanneswindelen.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
Markdown = "^3.3.4"
Jinja2 = "^2.11.3"
watchdog = "^2.0.2"
anytree = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.21.0"
black = "^20.8b1"
rope = "^0.18.0"
bs4 = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hyde = "hyde.hyde:cli"
